// client/src/types/VehicleMaster.ts
export default interface ITransportMaster {
  registration_no: string;
  chassi_no: string;
  engine_no: string;
  category: string;
  brand: string;
  color: string;
  remarks: string;
  registration_date: string;
  assign_pool: string;
  employee: string;
  insurance_company_name: string;
  insurance_expiry_date: string;
  license_expired_date: string;
  emission_test_date: string;
  availability: string;
}